+++ inst/extdata/zzz.R
@@ -1,0 +1,4 @@
+.onAttach <- function(libname, pkgname) {
+  dbVer = as.character(utils::packageVersion(pkgname))
+  packageStartupMessage('R ',pkgname,' Data v.',dbVer, '\n')
+}